foundry-extension/src/requirements.txt
fastapi
uvicorn
azure-identity
azure-ai-projects
